requirements.txt
asgiref==3.4.1
Django==3.2.23
django-filter==21.1
djangorestframework==3.14.0
PyMySQL==1.0.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.1.1
